Add a list of tasks for the Controller
